Improve BlockBuilder resizing behavior
Summary:
The most common pattern for BlockBuilder is to specify a max size, and then keep appending until it reports full before doing something with it. This is inherently inefficient because every time the underlying ByteArraySlice hits the full mark, it will resize and do a full copy of its contents. This means that that for most of the current use cases, we are double copying the data in the BlockBuilder. We can minimize this effect by using a slightly larger underlying storage size.