Fix bug in Benchmark data generation
Summary:
So I finally tracked down why so many of our benchmarks seemed to have taken such a huge performance hit.
It turns out there is was a bug in the data import code that caused the wrong column to be used which had
different performance characteristic. This change returns it back to the correct and previous behavior.
Local tests have shown that this is was the cause of the seeming perf degredation in our benchmark suite.