Prevent starvation when the task thread gets stuck processing a split
Because splits are scheduled by the task-processor on the shard-processor pool with "caller runs" rejection policy, it is possible
for the shard processors to finish their work and sit idle while the task processor finishes its current execution. As a result
the spare threads in the shared pool are not utilized efficiently and query latency gets affected.