Fix bug where fields may be written in the wrong order
This bug in CREATE TABLE AS SELECT was introduced by ad73cf8f57f70da2b165f0128c4c84ed6f5b2c14
with the removal of the call to unpack(...). This of this method was to turn multi-field
channels into one channel per field, but also reordered channels to match what the operator
expected.