Minor cleanup and (potential) optimization in compiler
Currently, the code generates calls to getCursor for every channel from 0 up to the
max channel referenced in filters/projections. With this change, the code now only
calls getCursor on channels that are actually needed for evaluation.

This is a re-do of 59e0cbabf465563c111350de9896e7019a81f01b (reverted in 87981e7bccfabacb001973193c4b7f120cc308a5)
with the bug fixed.