Skip warning when drop table cleanup fails
We try to cleanup any created tables in tests by dropping them, but the
drop fails if they don't exist, which usually means the table was never
created in the first place. Warning about this is confusing because the
warning is printed before the test failure and obscures the actual cause
of the test failure.