Fix property derivations for partitioned right or full joins
The derived properties of a hash-partitioned right or full join are currently
derived incorrectly from the properties of the probe side. Since this kind of join
can produce nulls from any of the partitions in case of a lack of matches,
the partitioning properties are violated for nulls.

This is problematic for queries that can benefit from the partitioning properties
after the join to avoid a reshuffle (e.g., an group by aggregation on the join key),
and may result in wrong results if the join produces any nulls on the probe side.

E.g.,

SELECT a.orderkey, count(*)
FROM (
   SELECT CASE WHEN orderkey > 2 THEN orderkey END orderkey
   FROM orders
   WHERE orderkey < 20
) a RIGHT OUTER JOIN (
   SELECT orderkey
   FROM orders
   WHERE orderkey < 20) b
ON a.orderkey = b.orderkey
GROUP BY 1;

Produces:

 orderkey | _col1
----------+-------
 NULL     |     1    <<<<<
 NULL     |     1    <<<<<
        4 |     1
        3 |     1
        7 |     1
        6 |     1
        5 |     1
(7 rows)