Give better error message when a column is not found
Do not assume a Dereference is a rowfield when it is not a column. Only assume it is a
rowfield when a column is found in its prefix. In the example below, the correct error
message should be 'Column t.y cannot be resolved' instead of 'Column t cannot be resolved'.
  SELECT t.y FROM (VALUES 1) t(x)