Print expanded frames for DUMP_BYTE_CODE_RAW
It is useful to have frame information when reading dumped byte code. Given
this line of code is for debug only, performance is not a concern. For ease
of reading, changing this option adds 1 line for every jump target in the
dumped byte code, which hardly affects reading if the reader doesn't need/want
this information.