Turn off optimizer for pushing table write through UNION ALL
We found a bug related to pushing table write through UNION ALL. We added a
configuration property and a session property to turn this optimizer on/off.
It is default to off so that users are not affected when we are fixing it.

I also added a test to exercises this bug. It fails when the optimizer is
turned on.