Add identifier to dictionary blocks
The identifier in a dictionary block represents the source block it was
projected from. If a dictionary block is projected into several blocks,
all the projected blocks will have the same identifier.

This is needed for deterministically corelating dictionary blocks in a
page for optimizations in operators and for compacting pages.