Fix stage/task leak
Given a query with three stages S1, S2, and S3 where S1 reads from both
S2 and S3. If S1 finishes reading from S2 before S2 finishes (for
example there's a LIMIT in the query), but the DELETE sent by
HttpPageBufferClient is lost for at least one task, S2 will not
transition to a terminal state until the entire query finishes. This
fixes this situation by cancelling all children of a stage when the stage
completes.